requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.6.1
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cytoolz==0.12.0
dacite==1.6.0
Deprecated==1.2.13
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Flask==2.2.2
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
install==1.3.5
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.3
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
parsimonious==0.8.1
Pillow==9.3.0
protobuf==3.20.3
pycryptodome==3.16.0
pyee==9.0.4
pymerkle==2.0.2
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
requests==2.28.1
rlp==2.0.1
six==1.16.0
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==3.0.0
toolz==0.12.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.27.0
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
